Read colonization files in.

Reads data

readData(infile, type = c("none", "trouvelot", "grid"))

Arguments

infile
the name of the file which the data are to be read from.

type
grid or trouvelot

Format

A grid dataset should have 7 variables:

It is mandatory that the dataset is formatted as follow
- Samples:

Sample names

- Replicates:

Sample Replicates

- Total:

Number of total fungal structures identified

- Hyphopodia:

Number of hypopodia identified

- IntrHyphae:

Number of interanl hyphae identified

- Arbuscules:

Number of arbuscules identified

- Vesicles:

Number of vesicles identified

A Trouvelot dataset should have 3 variables:

It is mandatory that the dataset is formatted as follow
- Samples:

Sample names

- Replicates:

Sample Replicates

- Scoring:

Scoring value. The only possible terms in this column are: 0A0,1A3,2A3,3A3,4A3,5A3,1A2,2A2,3A2,4A2,5A2,1A1,2A1, 3A1,4A1,5A1,1A0,2A0,3A0,4A0,5A0. If 0 is present, it will be converted into 0A0

The order of the Samples (in the "Samples" column) is the order used for plot display. So, if you like to have a specific order, please sort the original data accordingly

Value

an object of class grid or trouvelot.

Details

This funtion reads trouvelot and grid data

Examples

f <- dir(system.file("extdata", package = "Ramf"), full.names = TRUE, pattern = "grid.csv")
x <- readData(f, type = "grid")